// src/role/permissions.enum.ts
export enum Permission {
  STUDENT_MANAGEMENT = 'STUDENT_MANAGEMENT',
  TEACHER_MANAGEMENT = 'TEACHER_MANAGEMENT',
  ROLES_MANAGEMENT = 'ROLES_MANAGEMENT',
  SAVING_SESSION_MANAGEMENT = 'SAVING_SESSION_MANAGEMENT',
  ATTENDANCE_MANAGEMENT = 'ATTENDANCE_MANAGEMENT',
  CURRICULUM_MANAGEMENT = 'CURRICULUM_MANAGEMENT',
  SETTINGS_MANAGEMENT = 'SETTINGS_MANAGEMENT',
  POINTS_MANAGEMENT = 'POINTS_MANAGEMENT',
  AWARDS_MANAGEMENT = 'AWARDS_MANAGEMENT',
  ACCESS_DASHBOARD = 'ACCESS_DASHBOARD',
}